requirements.txt
streamlit==1.35.0
PyMuPDF==1.24.1
langchain==0.2.1
langchain-google-genai==1.0.2
python-dotenv==1.0.1
